# backend/docs/openapi/api.yaml
...
openapi: 3.0.3
info:
    description: |
        Machine-generated OpenAPI specification for AppOS APIs, with ext routes generated from route code using group-matrix as single source of truth.
    title: AppOS API
    version: 0.1.0
...
components:
    schemas:
        ErrorEnvelope:
            properties:
                code:
                    type: integer
                message:
                    type: string
            required:
                - message
            type: object
        GenericRequest:
            additionalProperties: true
            description: Generic request payload placeholder (refine per endpoint)
            type: object
...
        SuccessEnvelope:
            additionalProperties: true
            description: Generic success payload envelope
            example:
                ok: true
            type: object
    securitySchemes:
        bearerAuth:
            bearerFormat: PocketBase token
            scheme: bearer
            type: http
paths:
...
    /api/ext/services:
        get:
            description: Returns all supervisord programs with runtime status (RUNNING/STOPPED/etc.) and CPU/memory usage. Superuser only.
            operationId: get_api_ext_services
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessEnvelope'
                    description: OK
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "503":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Service Unavailable
            security:
                - bearerAuth: []
            summary: List services
            tags:
                - Services
    /api/ext/services/{name}/logs:
        get:
            description: Returns recent stdout or stderr log output for the named supervisord program. Superuser only.
            operationId: get_api_ext_services_name_logs
            parameters:
                - in: path
                  name: name
                  required: true
                  schema:
                    type: string
                - in: query
                  name: length
                  required: false
                  schema:
                    type: string
                - in: query
                  name: type
                  required: false
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessEnvelope'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
            summary: Get service logs
            tags:
                - Services
    /api/ext/services/{name}/restart:
        post:
            description: Restarts the named supervisord program. Writes an audit entry. Superuser only.
            operationId: post_api_ext_services_name_restart
            parameters:
                - in: path
                  name: name
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GenericRequest'
                required: false
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessEnvelope'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
            summary: Restart service
            tags:
                - Services
    /api/ext/services/{name}/start:
        post:
            description: Starts the named supervisord program. Already-running is treated as success. Superuser only.
            operationId: post_api_ext_services_name_start
            parameters:
                - in: path
                  name: name
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GenericRequest'
                required: false
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessEnvelope'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
            summary: Start service
            tags:
                - Services
    /api/ext/services/{name}/stop:
        post:
            description: Stops the named supervisord program. Already-stopped is treated as success. Superuser only.
            operationId: post_api_ext_services_name_stop
            parameters:
                - in: path
                  name: name
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GenericRequest'
                required: false
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessEnvelope'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
            summary: Stop service
            tags:
                - Services
...
    /api/ext/terminal/server/{serverId}/systemd/services:
        get:
            description: Returns the list of all systemd services and their statuses from the remote server. Superuser only.
            operationId: get_api_ext_terminal_server_serverid_systemd_services
            parameters:
                - in: path
                  name: serverId
                  required: true
                  schema:
                    type: string
                - in: query
                  name: keyword
                  required: false
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessEnvelope'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
            summary: List services
            tags:
                - Server Ops
...